SEQNG-244 Port of the log area #time 20m
