+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/SeqexecUI.scala
@@ -20,6 +20,7 @@ object SeqexecMain {
         NavBar(),
         QueueArea(SeqexecCircuit.statusAndLoadedSequences),
         SequenceArea(SeqexecCircuit.statusAndSequences, SeqexecCircuit.headerSideBarReader),
+        LogArea(SeqexecCircuit.log),
         LoginBox(SeqexecCircuit.loginBox)
       )
     ).build
