--- modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/sequence/LogArea.scala
@@ -1,36 +1,0 @@
-package edu.gemini.seqexec.web.client.components.sequence
-
-import diode.react.ModelProxy
-import edu.gemini.seqexec.web.client.model.GlobalLog
-import japgolly.scalajs.react.ScalaComponent
-import japgolly.scalajs.react.component.Scala.Unmounted
-import japgolly.scalajs.react.vdom.html_<^._
-
-/**
-  * Area to display a sequence's log
-  */
-object LogArea {
-  case class Props(log: GlobalLog)
-
-  private val component = ScalaComponent.builder[Props]("LogArea")
-    .stateless
-    .render_P(p =>
-      <.div(
-        ^.cls := "ui raised secondary segment",
-        <.h4("Log"),
-        <.div(
-          ^.cls := "ui form",
-          <.div(
-            ^.cls := "field",
-            <.textarea(
-              ^.readOnly := true,
-              ^.value := p.log.log.map(e => s"${e.timestamp} ${e.s}").mkString("\n")
-            )
-          )
-        )
-      )
-    )
-    .build
-
-  def apply(p: ModelProxy[GlobalLog]): Unmounted[Props, Unit, Unit] = component(Props(p()))
-}
